backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "Piyush Raj Mishra", email = "piyushxcoder@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = ["fastapi (>=0.128.0,<0.129.0)", "uvicorn (>=0.40.0,<0.41.0)", "pydantic-settings (>=2.12.0,<3.0.0)", "pydantic (>=2.12.5,<3.0.0)", "agno (>=2.4.3,<3.0.0)", "google-genai (>=1.60.0,<2.0.0)", "pandas (<3.0.0)", "python-multipart (>=0.0.22,<0.0.23)", "ollama (>=0.6.1,<0.7.0)", "openai (>=2.15.0,<3.0.0)", "mysql-connector-python (==9.5.0)"]

[tool.poetry]
packages = [{ include = "backend", from = "src" }]

[tool.poetry.scripts]
backend = "backend.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
